Crate transport

transport

Streams based communications with transport abstraction and NAT traversal.

Reexports

pub use protocol::Protocol;

pub use endpoint::{Endpoint, ToEndpoints};

pub use listen_endpoint::{ListenEndpoint, ToListenEndpoints};

pub use stream::{Stream, StreamInfo};

pub use listener::Listener;

Modules

endpoint
Defines the Endpoint type and related items.

listen_endpoint
Defines the ListenEndpoint type and related items.

listener
Defines the Listener type and related items.

protocol
Defines the protocols supported by this library through the Protocol type.

stream
Defines the Stream type and related items.

Structs

PrivRendezvousInfo
The private half of a rendezvous info pair. Used to perform rendezvous connections.

PubRendezvousInfo
The public half of a rendezvous info pair. Share this object with the remote peer and use their PubRendezvousInfo to perform a rendezvous connect.