Module transport::listener

Defines the Listener type and related items.

Structs

Listener
Listens on a set of listen endpoints and accepts incoming connections.

Enums

ListenerAcceptError
Errors returned by Listener::accept.

ListenerAddExternalEndpointsError
Errors returned by Listener::add_external_endpoints

ListenerBindEndpointError
Errors that can be returned when binding to a specific endpoint. Returned in ListenerBindError.

ListenerBindEndpointWarning
Warnings raised by Listener::bind