Enum transport::listener::ListenerAcceptError

pub enum ListenerAcceptError {
    TcpAccept {
        local_addr: SocketAddr,
        err: Error,
    },
    TcpSetTimeout {
        local_addr: SocketAddr,
        peer_addr: SocketAddr,
        err: Error,
    },
    TcpRead {
        local_addr: SocketAddr,
        peer_addr: SocketAddr,
        err: Error,
    },
    TcpWrite {
        local_addr: SocketAddr,
        peer_addr: SocketAddr,
        err: Error,
    },
    TcpStreamize {
        local_addr: SocketAddr,
        peer_addr: SocketAddr,
        err: StreamFromTcpStreamError,
    },
}

Errors returned by Listener::accept.

Variants

TcpAccept

Error accepting an incoming tcp connection.

Fields

local_addr

err

TcpSetTimeout

Error setting the timeout option on an incoming tcp stream.

Fields

local_addr

peer_addr

err

TcpRead

Error reading from freshly accepted tcp stream.

Fields

local_addr

peer_addr

err

TcpWrite

Error writing to freshly accepted tcp stream.

Fields

local_addr

peer_addr

err

TcpStreamize

Error building a Stream out of an accepted tcp stream.

Fields

local_addr

peer_addr

err