Struct transport::listener::Listener

pub struct Listener<M: AsRef<MappingContext> + Send + Clone> {
    // some fields omitted
}

Listens on a set of listen endpoints and accepts incoming connections.

Methods

impl<M> Listener<M> where M: AsRef<MappingContext> + Send + Clone

fn new(mc: M) -> Listener<M>

Create a new Listener which is not listening on any endpoints.

fn bind<E>(&self, listen_endpoints: E, deadline: Instant) -> Vec<WResult<ListenEndpoint, ListenerBindEndpointWarning, ListenerBindEndpointError<E::Err>>> where E: ToListenEndpoints

Bind to all the endpoints that can be parsed from listen_endpoints.

fn add_external_endpoints<E>(&self, listen_endpoint: ListenEndpoint, endpoints: E) -> Result<(), ListenerAddExternalEndpointsError<E::Err>> where E: ToEndpoints

Register an external endpoint with the Listener. This is for informing the Listener about endpoints that it could not discover itself using the MappingContext, such as manually configured external ports on a router.

fn external_endpoints(&self) -> Vec<Endpoint>

Retrieve a list of all our known external endpoints.

fn stop(&self)

Cause any concurrent call to Listener::accept to unblock and return None

fn accept(&self) -> Option<Result<Stream, ListenerAcceptError>>

Block until there is a new incoming connection or Listener::stop is called.

fn gen_rendezvous_info(&self, deadline: Instant) -> (PrivRendezvousInfo, PubRendezvousInfo, StreamGenRendezvousInfoDiagnostics)

Generate rendezvous info that can be used to connect to this listener set. The returned PrivRendezvousInfo should be used in a later call to Listener::rendezvous_connect. The PubRendezvousInfo should be swapped with that of the remote peer. The StreamGenRendezvousInfoDiagnostics contains information that may be useful in diagnosing problems with using this rendezvous info.

fn rendezvous_connect(&self, our_priv_info: PrivRendezvousInfo, their_pub_info: PubRendezvousInfo, deadline: Instant) -> WResult<Stream, StreamRendezvousConnectWarning, StreamRendezvousConnectError>

Perform a rendezvous connection, making use of the listeners in this listener set. This function should be called by both peers wishing to connect. Both peers will attempt to connect to each others listeners while also attempting to make nat-traversed connections.