Enum transport::stream::StreamRendezvousConnectError

pub enum StreamRendezvousConnectError {
    Expired,
    AllProtocolsFailed {
        tcp_err: Option<StreamRendezvousConnectTcpError>,
        direct_err: Option<StreamDirectConnectError<Void>>,
    },
}

Errors returned by Stream::rendezvous_connect.

Variants

Expired

The supplied rendezvous info has expired.

AllProtocolsFailed

Failed to make a rendezvous connection with any protocol.

Fields

tcp_err

The error of the failed tcp rendezvous connect attempt (if any).

direct_err

The error (if any) of the failed attempt to directly connect to the peer's listeners.